+++ src/simulation/config.rs
@@ -1,4 +1,4 @@
-use uom::si::f32::Time;
+use uom::si::f64::Time;
 
 use crate::simulation::units::*;
 
@@ -6,5 +6,5 @@ pub struct SimulationConfig {
     pub time_step: Time,
     pub length_unit: LengthUnit,
     pub mass_unit: MassUnit,
-    pub pixels_per_length: f32,
+    pub pixels_per_length: f64,
 }
